requirements.txt
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.3.1
Flask-Script==2.0.6
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
psycopg2-binary==2.7.7
numpy==1.16.1
pandas==0.24.1
SQLAlchemy==1.2.17
Werkzeug==0.14.1
Flask-Cors==3.0.7
python-dotenv==0.10.1
pytest==4.2.0
coverage==4.5.2
waitress==1.2.1
